refactor: my-page, main-page
