+++ src/main/java/com/bovo/Bovo/modules/main/controller/MainController.java
@@ -1,9 +1,8 @@
 package com.bovo.Bovo.modules.main.controller;
 
-import com.bovo.Bovo.modules.main.dto.response.BookListDto;
-import com.bovo.Bovo.modules.main.dto.response.RecentBookInfoDto;
+import com.bovo.Bovo.modules.main.dto.response.partial.RecentBookInfoDto;
 import com.bovo.Bovo.modules.main.dto.response.TotalMainDto;
-import com.bovo.Bovo.modules.main.dto.response.UserInfoDto;
+import com.bovo.Bovo.modules.main.dto.response.partial.UserInfoDto;
 import com.bovo.Bovo.modules.main.service.MainService;
 import com.bovo.Bovo.modules.my_page.service.MyPageService;
 import com.bovo.Bovo.modules.user.dto.security.AuthenticatedUserId;
@@ -14,6 +13,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.Map;
 
 @Slf4j
 @RestController
@@ -29,11 +30,15 @@ public class MainController {
         UserInfoDto userInfoDto = mainService.getUserInfoByUserId(userId);
         int totalBookNum = myPageService.countCompletedBooksByUserId(userId);
         RecentBookInfoDto recentBookInfoDto = mainService.getRecentBookInfoByUserId(userId);
-        BookListDto bookListDto = mainService.getBookListByUserId(userId);
+        Map<String, String> bookList = mainService.getBookListByUserId(userId);
 
         return ResponseEntity.status(HttpStatus.OK)
                 .body(new TotalMainDto(200, "메인 페이지 로드 완료",
                         userInfoDto.getProfile_picture(), userInfoDto.getNickname(), userInfoDto.getLevel(), totalBookNum,
-                        recentBookInfoDto, bookListDto));
+                        recentBookInfoDto, bookList));
+//        return ResponseEntity.status(HttpStatus.OK)
+//                .body(new TotalMainDto(200, "메인 페이지 로드 완료",
+//                        userInfoDto.getProfile_picture(), userInfoDto.getNickname(), userInfoDto.getLevel(), 30,
+//                        recentBookInfoDto, bookList));
     }
 }
