--- src/main/java/com/bovo/Bovo/modules/main/dto/response/BookListDto.java
@@ -1,14 +1,0 @@
-package com.bovo.Bovo.modules.main.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.Map;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class BookListDto {
-    private Map<String, String> bookList;
-}
