+++ src/main/java/com/bovo/Bovo/modules/main/service/MainService.java
@@ -1,13 +1,14 @@
 package com.bovo.Bovo.modules.main.service;
 
-import com.bovo.Bovo.modules.main.dto.response.BookListDto;
-import com.bovo.Bovo.modules.main.dto.response.RecentBookInfoDto;
-import com.bovo.Bovo.modules.main.dto.response.UserInfoDto;
+import com.bovo.Bovo.modules.main.dto.response.partial.RecentBookInfoDto;
+import com.bovo.Bovo.modules.main.dto.response.partial.UserInfoDto;
+
+import java.util.Map;
 
 public interface MainService {
     UserInfoDto getUserInfoByUserId(Integer userId);
 
     RecentBookInfoDto getRecentBookInfoByUserId(Integer userId);
 
-    BookListDto getBookListByUserId(Integer userId);
+    Map<String, String> getBookListByUserId(Integer userId);
 }
